requirements.txt
streamlit
pandas
yfinance
matplotlib
scikit-learn
requests
python-dotenv